Decouple Transport components and Networking.SemanticData #321
- added new tests to MessageWriterBaseTest.cs and rearranged related stuff
- UT :-1: some test fails
