+++ Networking/Tests/SemanticDatalUnitTest/BinaryDecoderUnitTest.cs
@@ -7,7 +7,6 @@
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
-using System.Collections.Generic;
 using System.Globalization;
 using UAOOI.Configuration.Networking.Serialization;
 using UAOOI.Networking.Core;
